refactor test utils into seperate file
